requirements.txt
# --- Core do Backend e Servidor ---
# Framework web, servidor ASGI e servidor de producao para Heroku.
fastapi
uvicorn
gunicorn
python-dotenv

# --- Banco de Dados ---
# ORM para interagir com o PostgreSQL e o driver necessario. Alembic para migracoes.
sqlalchemy
psycopg2-binary
alembic

# --- Validacao e Processamento de Dados ---
# Pydantic e essencial para o FastAPI. Orjson para performance com JSON.
pydantic
orjson

# --- IA: LLM e LangChain ---
# Bibliotecas para interagir com a API do Google Gemini e orquestrar os prompts.
google-generativeai
langchain-core
langchain-google-genai
requests

numpy
scipy
scikit-learn
openai

# --- Orquestracao Agentica (v2.0) ---
# LangGraph para pipeline agentico com grafo de estados.
langgraph

# --- Vetores e Busca Semantica (v2.0) ---
# pgvector para embeddings nativos no PostgreSQL.
pgvector

# --- Processamento de Arquivos ---
# A documentacao cita explicitamente o uso de PyMuPDF para extracao de dados de PDFs.
PyMuPDF
PyPDF2

# --- Autenticacao e Seguranca (v3.0 Multi-Tenant) ---
# JWT para tokens de sessao e passlib para hash de senhas.
python-jose[cryptography]
passlib[bcrypt]

# --- Utilitarios ---
# 'rich' para logs e saidas formatadas no console, o que e util para debug dos scripts.
# 'APScheduler' para tarefas agendadas, como descrito na tabela 'configuracoes_coleta'.
rich
APScheduler
python-multipart
